refactor: move facets code from search to filters.js
