+++ src/components/Search.js
@@ -35,6 +35,7 @@ import {
     shopifyDefaultFields,
 } from '../utils';
 import GeoResultsLayout from './GeoLayout/GeoResultsLayout';
+import Filters from './Filters';
 
 const { Meta } = Card;
 const { Panel } = Collapse;
@@ -428,740 +429,6 @@ class Search extends Component {
         return fontFamily ? { fontFamily } : {};
     };
 
-    renderCollectionFilter = (font) => {
-        if (!this.collectionFilter) {
-            return null;
-        }
-
-        const type = get(this.collectionFilter, 'rsConfig.filterType', '');
-        if(type === 'list') {
-            return (
-                <React.Fragment>
-                    <ReactiveComponent
-                        componentId="filter_by_collection"
-                        customQuery={() =>
-                            this.exportType === 'shopify'
-                                ? {
-                                      query: {
-                                          term: { 'type.keyword': ['collection'] },
-                                      },
-                                  }
-                                : null
-                        }
-                    />
-                    <MultiList
-                        innerClass={{
-                            input: 'list-input'
-                        }}
-                        componentId="collection"
-                        dataField="collection"
-                        style={{
-
-                        }}
-                        defaultQuery={() => ({
-                            aggs: {
-                                collections: {
-                                    terms: {
-                                        field: '_id',
-                                        size: 50,
-                                        order: {
-                                            'product_count.value': 'desc',
-                                        },
-                                    },
-                                    aggs: {
-                                        top_collections: {
-                                            top_hits: {
-                                                _source: {
-                                                    includes: [
-                                                        'title',
-                                                        'product_count',
-                                                    ],
-                                                },
-                                                size: 1,
-                                            },
-                                        },
-                                        product_count: {
-                                            sum: {
-                                                field: 'product_count',
-                                            },
-                                        },
-                                    },
-                                },
-                            },
-                        })}
-                        size={50}
-                        showCheckbox={this.themeType !== 'minimal'}
-                        react={{
-                            and: [
-                                'filter_by_collection',
-                                // TODO: Make it reactive to other filters
-                                // ...getReactDependenciesFromPreferences(
-                                //     this.preferences,
-                                //     'collection',
-                                // ),
-                            ],
-                        }}
-                        // TODO: transform the value to title later
-                        showFilter={false}
-                        render={({ loading, data, value, handleChange }) => {
-                            if (loading) {
-                                return (
-                                    <div
-                                        css={loaderStyle}
-                                        // eslint-disable-next-line
-                                        dangerouslySetInnerHTML={{
-                                            __html: get(
-                                                this.collectionFilter,
-                                                'customMessages.loading',
-                                                'Loading collections',
-                                            ),
-                                        }}
-                                    />
-                                );
-                            }
-                            return (
-                                <UL role="listbox" aria-label="collection-items">
-                                    {data.length ? null : (
-                                        <div
-                                            // eslint-disable-next-line
-                                            dangerouslySetInnerHTML={{
-                                                __html: get(
-                                                    this.collectionFilter,
-                                                    'customMessages.noResults',
-                                                    'No items Found',
-                                                ),
-                                            }}
-                                        />
-                                    )}
-                                    {data.map((item) => {
-                                        const isChecked = !!value[item.key];
-                                        const title = get(
-                                            item,
-                                            'top_collections.hits.hits[0]._source.title',
-                                        );
-                                        const count = get(
-                                            item,
-                                            'top_collections.hits.hits[0]._source.product_count',
-                                        );
-                                        return (
-                                            <li
-                                                key={item.key}
-                                                className={`${
-                                                    isChecked ? 'active' : ''
-                                                }`}
-                                                role="option"
-                                                aria-checked={isChecked}
-                                                aria-selected={isChecked}
-                                            >
-                                                <Checkbox
-                                                    id={`collection-${item.key}`}
-                                                    name={`collection-${item.key}`}
-                                                    value={item.key}
-                                                    onChange={handleChange}
-                                                    checked={isChecked}
-                                                    show
-                                                />
-                                                {/* eslint-disable-next-line */}
-                                                <label
-                                                    htmlFor={`collection-${item.key}`}
-                                                >
-                                                    <span>
-                                                        <span>{title}</span>
-                                                        <span>{count}</span>
-                                                    </span>
-                                                </label>
-                                            </li>
-                                        );
-                                    })}
-                                </UL>
-                            );
-                        }}
-                        URLParams
-                        {...get(this.collectionFilter, 'rsConfig')}
-                        title=""
-                    />
-                </React.Fragment>
-            );
-        }
-
-        if( get(this.collectionFilter, 'rsConfig.startValue', '') && get(this.collectionFilter, 'rsConfig.endValue', '')) {
-            return (
-                <RangeInput
-                    key="filter_by_collection"
-                    componentId="filter_by_collection"
-                    dataField={get(
-                        this.collectionFilter,
-                        'rsConfig.dataField',
-                        shopifyDefaultFields.size,
-                    )}
-                    range={{
-                        start: parseInt(get(
-                            this.collectionFilter,
-                            'rsConfig.startValue',
-                            ''
-                        ), 10),
-                        end: parseInt(get(
-                            this.collectionFilter,
-                            'rsConfig.endValue',
-                            ''
-                        ), 10),
-                    }}
-                    rangeLabels={{
-                        start: get(
-                            this.collectionFilter,
-                            'rsConfig.startLabel',
-                            ''
-                        ),
-                        end: get(
-                            this.collectionFilter,
-                            'rsConfig.endLabel',
-                            ''
-                        ),
-                    }}
-                    showHistogram={get(
-                        this.collectionFilter,
-                        'rsConfig.showHistogram',
-                        false
-                    )}
-                    filterLabel={get(this.collectionFilter, 'rsConfig.title', 'Collection')}
-                    URLParams
-                    css={font}
-                />
-            );
-        }
-        return (
-            <DynamicRangeSlider
-                key="filter_by_collection"
-                componentId="filter_by_collection"
-                dataField={get(
-                    this.collectionFilter,
-                    'rsConfig.dataField',
-                    shopifyDefaultFields.size,
-                )}
-                showHistogram={get(
-                    this.collectionFilter,
-                    'rsConfig.showHistogram',
-                    false
-                )}
-                filterLabel={get(this.collectionFilter, 'rsConfig.title', 'Collection')}
-                URLParams
-                css={font}
-            />
-        )
-
-    };
-
-    renderProductTypeFilter = (font) => {
-        if (!this.productTypeFilter) {
-            return null;
-        }
-
-        const type = get(this.productTypeFilter, 'rsConfig.filterType', '');
-        if(type === 'list') {
-            return (
-                <MultiList
-                    componentId="productType"
-                    dataField="product_type.keyword"
-                    innerClass={{
-                        input: 'list-input'
-                    }}
-                    css={font}
-                    showCheckbox={this.themeType !== 'minimal'}
-                    react={{
-                        and: [
-                            ...getReactDependenciesFromPreferences(
-                                this.preferences,
-                                'productType',
-                            ),
-                        ],
-                    }}
-                    filterLabel="Product Type"
-                    URLParams
-                    {...get(this.productTypeFilter, 'rsConfig')}
-                    title=""
-                />
-            );
-        }
-
-        if( get(this.productTypeFilter, 'rsConfig.startValue', '') && get(this.productTypeFilter, 'rsConfig.endValue', '')) {
-            return (
-                <RangeInput
-                    key="productType"
-                    componentId="productType"
-                    dataField="product_type"
-                    range={{
-                        start: parseInt(get(
-                            this.productTypeFilter,
-                            'rsConfig.startValue',
-                            ''
-                        ), 10),
-                        end: parseInt(get(
-                            this.productTypeFilter,
-                            'rsConfig.endValue',
-                            ''
-                        ), 10),
-                    }}
-                    rangeLabels={{
-                        start: get(
-                            this.productTypeFilter,
-                            'rsConfig.startLabel',
-                            ''
-                        ),
-                        end: get(
-                            this.productTypeFilter,
-                            'rsConfig.endLabel',
-                            ''
-                        ),
-                    }}
-                    showHistogram={get(
-                        this.productTypeFilter,
-                        'rsConfig.showHistogram',
-                        false
-                    )}
-                    URLParams
-                    filterLabel={get(this.productTypeFilter, 'rsConfig.title', 'productType')}
-                    css={font}
-                />
-            );
-        }
-        return (
-            <DynamicRangeSlider
-                key="productType"
-                componentId="productType"
-                dataField={get(
-                    this.productTypeFilter,
-                    'rsConfig.dataField',
-                    shopifyDefaultFields.size,
-                )}
-                showHistogram={get(
-                    this.productTypeFilter,
-                    'rsConfig.showHistogram',
-                    false
-                )}
-                URLParams
-                css={font}
-                filterLabel={get(this.productTypeFilter, 'rsConfig.title', 'productType')}
-            />
-        )
-    };
-
-    renderColorFilter = (font) => {
-        const type = get(this.colorFilter, 'rsConfig.filterType', '');
-        if(type === 'list') {
-            return (
-                <MultiList
-                    componentId="color"
-                    innerClass={{
-                        input: 'list-input'
-                    }}
-                    react={{
-                        and: [
-                            'colorOption',
-                            ...getReactDependenciesFromPreferences(
-                                this.preferences,
-                                'color',
-                            ),
-                        ],
-                    }}
-                    showSearch={false}
-                    css={font}
-                    showCheckbox={this.themeType !== 'minimal'}
-                    render={({ loading, error, data, handleChange, value }) => {
-                        const values = [...new Set(Object.keys(value))];
-                        const browserStringColors = Object.keys(browserColors);
-                        if (loading) {
-                            return (
-                                <div
-                                    css={loaderStyle}
-                                    // eslint-disable-next-line
-                                    dangerouslySetInnerHTML={{
-                                        __html: get(
-                                            this.colorFilter,
-                                            'customMessages.noResults',
-                                            'Fetching Colors',
-                                        ),
-                                    }}
-                                />
-                            );
-                        }
-                        if (error) {
-                            return (
-                                <div>
-                                    No colors found!
-                                </div>
-                            );
-                        }
-                        if (data.length === 0) {
-                            return (
-                                <div
-                                    // eslint-disable-next-line
-                                    dangerouslySetInnerHTML={{
-                                        __html: get(
-                                            this.colorFilter,
-                                            'customMessages.noResults',
-                                            'Fetching Colors',
-                                        ),
-                                    }}
-                                />
-                            );
-                        }
-                        const primaryColor =
-                            get(this.theme, 'colors.primaryColor', '') || '#0B6AFF';
-                        const normalizedData = [];
-                        data.forEach((i) => {
-                            if (
-                                !normalizedData.find(
-                                    (n) => n.key === i.key.toLowerCase(),
-                                )
-                            ) {
-                                normalizedData.push({
-                                    ...i,
-                                    key: i.key.toLowerCase(),
-                                });
-                            }
-                        });
-                        return (
-                            <div css={colorContainer}>
-                                {normalizedData.map((item) =>
-                                    browserStringColors.includes(
-                                        item.key.toLowerCase(),
-                                    ) ? (
-                                        <Tooltip
-                                            key={item.key}
-                                            placement="top"
-                                            title={item.key}
-                                        >
-                                            {/* eslint-disable-next-line */}
-                                            <div
-                                                onClick={() => handleChange(item.key)}
-                                                style={{
-                                                    width: '100%',
-                                                    height: 30,
-                                                    background: item.key,
-                                                    transition: 'all ease .2s',
-                                                    cursor: 'pointer',
-                                                    border:
-                                                        values &&
-                                                        values.includes(item.key)
-                                                            ? `2px solid ${primaryColor}`
-                                                            : `1px solid #ccc`,
-                                                }}
-                                            />
-                                        </Tooltip>
-                                    ) : null,
-                                )}
-                            </div>
-                        );
-                    }}
-                    loader={
-                        <div
-                            css={loaderStyle}
-                            // eslint-disable-next-line
-                            dangerouslySetInnerHTML={{
-                                __html: get(
-                                    this.colorFilter,
-                                    'customMessages.loading',
-                                    'Loading colors',
-                                ),
-                            }}
-                        />
-                    }
-                    URLParams
-                    {...get(this.colorFilter, 'rsConfig')}
-                    dataField={get(
-                        this.colorFilter,
-                        'rsConfig.dataField',
-                        shopifyDefaultFields.color,
-                    )}
-                    title=""
-                />
-            );
-        }
-
-        if( get(this.colorFilter, 'rsConfig.startValue', '') && get(this.colorFilter, 'rsConfig.endValue', '')) {
-            return (
-                <RangeInput
-                    key="color"
-                    componentId="color"
-                    dataField={get(
-                        this.colorFilter,
-                        'rsConfig.dataField',
-                        shopifyDefaultFields.size,
-                    )}
-                    range={{
-                        start: parseInt(get(
-                            this.colorFilter,
-                            'rsConfig.startValue',
-                            ''
-                        ), 10),
-                        end: parseInt(get(
-                            this.colorFilter,
-                            'rsConfig.endValue',
-                            ''
-                        ), 10),
-                    }}
-                    rangeLabels={{
-                        start: get(
-                            this.colorFilter,
-                            'rsConfig.startLabel',
-                            ''
-                        ),
-                        end: get(
-                            this.colorFilter,
-                            'rsConfig.endLabel',
-                            ''
-                        ),
-                    }}
-                    showHistogram={get(
-                        this.colorFilter,
-                        'rsConfig.showHistogram',
-                        false
-                    )}
-                    URLParams
-                    css={font}
-                    filterLabel={get(this.colorFilter, 'rsConfig.title', 'color')}
-                />
-            )
-        }
-        return (
-            <DynamicRangeSlider
-                key="color"
-                componentId="color"
-                dataField={get(
-                    this.colorFilter,
-                    'rsConfig.dataField',
-                    shopifyDefaultFields.size,
-                )}
-                showHistogram={get(
-                    this.colorFilter,
-                    'rsConfig.showHistogram',
-                    false
-                )}
-                URLParams
-                filterLabel={get(this.colorFilter, 'rsConfig.title', 'color')}
-                css={font}
-            />
-        )
-
-    }
-
-    renderSizeFilter = (font) => {
-        const type = get(this.sizeFilter, 'rsConfig.filterType', '');
-        if(type === 'list') {
-            return (
-                <React.Fragment>
-                    <MultiList
-                        componentId="size"
-                        innerClass={{
-                            input: 'list-input'
-                        }}
-                        react={{
-                            and: [
-                                'sizeOption',
-                                ...getReactDependenciesFromPreferences(
-                                    this.preferences,
-                                    'size',
-                                ),
-                            ],
-                        }}
-                        css={font}
-                        loader={
-                            <div
-                                css={loaderStyle}
-                                // eslint-disable-next-line
-                                dangerouslySetInnerHTML={{
-                                    __html: get(
-                                        this.sizeFilter,
-                                        'customMessages.loading',
-                                        'Loading sizes',
-                                    ),
-                                }}
-                            />
-                        }
-                        renderNoResults={() => (
-                            <div
-                                // eslint-disable-next-line
-                                dangerouslySetInnerHTML={{
-                                    __html: get(
-                                        this.sizeFilter,
-                                        'customMessages.noResults',
-                                        'No sizes Found',
-                                    ),
-                                }}
-                            />
-                        )}
-                        showCheckbox={this.themeType !== 'minimal'}
-                        URLParams
-                        {...get(this.sizeFilter, 'rsConfig')}
-                        dataField={get(
-                            this.sizeFilter,
-                            'rsConfig.dataField',
-                            shopifyDefaultFields.size,
-                        )}
-                        title=""
-                    />
-                </React.Fragment>
-            );
-        }
-
-        if( get(this.sizeFilter, 'rsConfig.startValue', '') && get(this.sizeFilter, 'rsConfig.endValue', '')) {
-            return (
-                <RangeInput
-                    key="size"
-                    componentId="size"
-                    dataField={get(
-                        this.sizeFilter,
-                        'rsConfig.dataField',
-                        shopifyDefaultFields.size,
-                    )}
-                    range={{
-                        start: parseInt(get(
-                            this.sizeFilter,
-                            'rsConfig.startValue',
-                            ''
-                        ), 10),
-                        end: parseInt(get(
-                            this.sizeFilter,
-                            'rsConfig.endValue',
-                            ''
-                        ), 10),
-                    }}
-                    rangeLabels={{
-                        start: get(
-                            this.sizeFilter,
-                            'rsConfig.startLabel',
-                            ''
-                        ),
-                        end: get(
-                            this.sizeFilter,
-                            'rsConfig.endLabel',
-                            ''
-                        ),
-                    }}
-                    showHistogram={get(
-                        this.sizeFilter,
-                        'rsConfig.showHistogram',
-                        false
-                    )}
-                    URLParams
-                    css={font}
-                    filterLabel={get(this.sizeFilter, 'rsConfig.title', 'size')}
-                />
-            );
-        }
-        return (
-            <DynamicRangeSlider
-                key="size"
-                componentId="size"
-                dataField={get(
-                    this.sizeFilter,
-                    'rsConfig.dataField',
-                    shopifyDefaultFields.size,
-                )}
-                showHistogram={get(
-                    this.sizeFilter,
-                    'rsConfig.showHistogram',
-                    false
-                )}
-                URLParams
-                css={font}
-                filterLabel={get(this.sizeFilter, 'rsConfig.title', 'size')}
-            />
-        )
-
-    }
-
-    renderPriceFilter = (font) => {
-        if( get(this.priceFilter, 'rsConfig.startValue', '') && get(this.priceFilter, 'rsConfig.endValue', '')) {
-            return (
-                <RangeInput
-                    componentId="price"
-                    dataField={get(
-                        this.priceFilter,
-                        'rsConfig.dataField',
-                        'variants.price',
-                    )}
-                    range={{
-                        start: parseInt(get(
-                            this.priceFilter,
-                            'rsConfig.startValue',
-                            ''
-                        ), 10),
-                        end: parseInt(get(
-                            this.priceFilter,
-                            'rsConfig.endValue',
-                            ''
-                        ), 10),
-                    }}
-                    rangeLabels={{
-                        start: get(
-                            this.priceFilter,
-                            'rsConfig.startLabel',
-                            ''
-                        ),
-                        end: get(
-                            this.priceFilter,
-                            'rsConfig.endLabel',
-                            ''
-                        ),
-                    }}
-                    showHistogram={get(
-                        this.priceFilter,
-                        'rsConfig.showHistogram',
-                        false
-                    )}
-                    URLParams
-                    css={font}
-                    filterLabel={get(this.priceFilter, 'rsConfig.title', 'size')}
-                />
-            )
-        }
-
-        return (
-            <DynamicRangeSlider
-                componentId="price"
-                dataField={get(
-                    this.priceFilter,
-                    'rsConfig.dataField',
-                    'variants.price',
-                )}
-                showHistogram={get(
-                    this.priceFilter,
-                    'rsConfig.showHistogram',
-                    false
-                )}
-                URLParams
-                css={font}
-                style={{
-                    marginTop: 50,
-                }}
-                loader={
-                    <div
-                        css={loaderStyle}
-                        // eslint-disable-next-line
-                        dangerouslySetInnerHTML={{
-                            __html: get(
-                                this.priceFilter,
-                                'customMessages.loading',
-                                '',
-                            ),
-                        }}
-                    />
-                }
-                rangeLabels={(min, max) => ({
-                    start: `${
-                        this.currency
-                    } ${min.toFixed(2)}`,
-                    end: `${
-                        this.currency
-                    } ${max.toFixed(2)}`,
-                })}
-                {...this.priceFilter.rsConfig}
-                title=""
-            />
-        )
-    }
-
     renderCategorySearch = (categorySearchProps) => {
         const { toggleFilters, blur } = this.state;
         const { isPreview } = this.props;
@@ -1240,20 +507,6 @@ class Search extends Component {
                 highlight={get(this.resultSettings, 'resultHighlight', false)}
             />
         );
-    };
-
-    showCollapseFilters = (componentsIdArray) => {
-        const {
-            settings: { isFilterCollapsible },
-            isMobile,
-        } = this.state;
-        if (isMobile) {
-            return componentsIdArray;
-        }
-        if (isFilterCollapsible) {
-            return [];
-        }
-        return componentsIdArray;
     };
 
     render() {
@@ -1366,436 +619,18 @@ class Search extends Component {
                             gridGap: 20,
                         }}
                     >
-                        <div
-                            css={{
-                                display: 'grid',
-                                gridTemplateColumns:
-                                    'repeat(auto-fit, minmax (250px, 1fr))',
-                                gridGap: 0,
-                                alignSelf: 'start',
-                                border:
-                                    this.themeType === 'classic'
-                                        ? '1px solid #eee'
-                                        : 0,
-                                [mediaMax.medium]: {
-                                    display: toggleFilters ? 'grid' : 'none',
-                                    gridTemplateColumns: '1fr',
-                                },
-                                boxShadow:
-                                    this.themeType === 'minimal'
-                                        ? `0 0 4px ${get(
-                                              this.theme,
-                                              'colors.titleColor',
-                                          )}1a`
-                                        : 0,
-                                [mediaMax.medium]: {
-                                    display: toggleFilters ? 'grid' : 'none',
-                                },
-                            }}
-                        >
-                            <Collapse
-                                bordered={false}
-                                defaultActiveKey={getReactDependenciesFromPreferences(
-                                    this.preferences,
-                                )}
-                            >
-                                {this.productTypeFilter ? (
-                                    <Panel
-                                        header={
-                                            <span
-                                                style={{
-                                                    color: get(
-                                                        this.theme,
-                                                        'colors.titleColor',
-                                                    ),
-                                                    fontWeight: 'bold',
-                                                    fontSize: '15px',
-                                                }}
-                                            >
-                                                {get(
-                                                    this.productTypeFilter,
-                                                    'rsConfig.title',
-                                                    'Product Type',
-                                                )}
-                                            </span>
-                                        }
-                                        showArrow={this.themeType !== 'minimal'}
-                                        key="productType"
-                                        css={this.getFontFamily()}
-                                        className="filter"
-                                    >
-                                        {this.renderProductTypeFilter(
-                                            this.getFontFamily,
-                                        )}
-                                    </Panel>
-                                ) : null}
-                                {this.collectionFilter ? (
-                                    <Panel
-                                        header={
-                                            <span
-                                                style={{
-                                                    color: get(
-                                                        this.theme,
-                                                        'colors.titleColor',
-                                                    ),
-                                                    fontWeight: 'bold',
-                                                    fontSize: '15px',
-                                                }}
-                                            >
-                                                {get(
-                                                    this.collectionFilter,
-                                                    'rsConfig.title',
-                                                    'Collections',
-                                                )}
-                                            </span>
-                                        }
-                                        showArrow={this.themeType !== 'minimal'}
-                                        key="collection"
-                                        css={this.getFontFamily()}
-                                        className="filter"
-                                    >
-                                        {this.renderCollectionFilter(
-                                            this.getFontFamily,
-                                        )}
-                                    </Panel>
-                                ) : null}
-                                {this.colorFilter ? (
-                                    <Panel
-                                        className="filter"
-                                        header={
-                                            <span
-                                                style={{
-                                                    color: get(
-                                                        this.theme,
-                                                        'colors.titleColor',
-                                                    ),
-                                                    fontWeight: 'bold',
-                                                    fontSize: '15px',
-                                                }}
-                                            >
-                                                {get(
-                                                    this.colorFilter,
-                                                    'rsConfig.title',
-                                                    'Color',
-                                                )}
-                                            </span>
-                                        }
-                                        showArrow={this.themeType !== 'minimal'}
-                                        key="color"
-                                        css={this.getFontFamily()}
-                                    >
-                                        {this.renderColorFilter(
-                                            this.getFontFamily,
-                                        )}
-                                    </Panel>
-                                ) : null}
-                                {this.sizeFilter ? (
-                                    <Panel
-                                        className="filter"
-                                        header={
-                                            <span
-                                                style={{
-                                                    color: get(
-                                                        this.theme,
-                                                        'colors.titleColor',
-                                                    ),
-                                                    fontWeight: 'bold',
-                                                    fontSize: '15px',
-                                                }}
-                                            >
-                                                {get(
-                                                    this.sizeFilter,
-                                                    'rsConfig.title',
-                                                    'Size',
-                                                )}
-                                            </span>
-                                        }
-                                        showArrow={this.themeType !== 'minimal'}
-                                        key="size"
-                                        css={this.getFontFamily()}
-                                    >
-                                        {this.renderSizeFilter(
-                                            this.getFontFamily(),
-                                        )}
-                                    </Panel>
-                                ) : null}
-                                {this.priceFilter ? (
-                                    <Panel
-                                        header={
-                                            <span
-                                                style={{
-                                                    color: get(
-                                                        this.theme,
-                                                        'colors.titleColor',
-                                                    ),
-                                                    fontWeight: 'bold',
-                                                    fontSize: '15px',
-                                                }}
-                                            >
-                                                {get(
-                                                    this.priceFilter,
-                                                    'rsConfig.title',
-                                                    'Price',
-                                                )}
-                                            </span>
-                                        }
-                                        showArrow={this.themeType !== 'minimal'}
-                                        key="price"
-                                        css={this.getFontFamily()}
-                                        className="filter"
-                                    >
-                                        {this.renderPriceFilter(
-                                            this.getFontFamily(),
-                                        )}
-                                        {/* <DynamicRangeSlider
-                                            componentId="price"
-                                            dataField="retail_price"
-                                            showHistogram={false}
-                                            css={this.getFontFamily()}
-                                            style={{
-                                                marginTop: 50,
-                                            }}
-                                            loader={
-                                                <div
-                                                    css={loaderStyle}
-                                                    // eslint-disable-next-line
-                                                    dangerouslySetInnerHTML={{
-                                                        __html: get(
-                                                            this.priceFilter,
-                                                            'customMessages.loading',
-                                                            '',
-                                                        ),
-                                                    }}
-                                                />
-                                            }
-                                            rangeLabels={(min, max) => ({
-                                                start: `${
-                                                    this.currency
-                                                } ${min.toFixed(2)}`,
-                                                end: `${
-                                                    this.currency
-                                                } ${max.toFixed(2)}`,
-                                            })}
-                                            // react={{
-                                            //     and: getReactDependenciesFromPreferences(
-                                            //         this.preferences,
-                                            //         'price',
-                                            //     ),
-                                            // }}
-                                            {...this.priceFilter.rsConfig}
-                                            title=""
-                                        /> */}
-                                    </Panel>
-                                ) : null}
-                                {this.dynamicFacets.map((listComponent) => (
-                                    <Panel
-                                        header={
-                                            <span
-                                                style={{
-                                                    color: get(
-                                                        this.theme,
-                                                        'colors.titleColor',
-                                                    ),
-                                                    fontWeight: 'bold',
-                                                    fontSize: '15px',
-                                                }}
-                                            >
-                                                {get(
-                                                    listComponent,
-                                                    'rsConfig.title',
-                                                )}
-                                            </span>
-                                        }
-                                        showArrow={this.themeType !== 'minimal'}
-                                        key={get(
-                                            listComponent,
-                                            'rsConfig.componentId',
-                                        )}
-                                        css={{
-                                            ...this.getFontFamily(),
-                                            maxWidth: isMobile
-                                                ? 'none'
-                                                : '298px',
-                                        }}
-                                        className="filter"
-                                    >
-                                        {
-                                            // eslint-disable-next-line no-nested-ternary
-                                            listComponent?.rsConfig?.filterType === 'range' ? (
-                                                (listComponent?.rsConfig?.startValue && listComponent?.rsConfig?.endValue) ? (
-                                                    <RangeInput
-                                                        key={get(
-                                                            listComponent,
-                                                            'rsConfig.componentId',
-                                                            ''
-                                                        )}
-                                                        componentId={get(
-                                                            listComponent,
-                                                            'rsConfig.componentId',
-                                                            ''
-                                                        )}
-                                                        dataField={get(
-                                                            listComponent,
-                                                            'rsConfig.dataField',
-                                                            ''
-                                                        )}
-                                                        range={{
-                                                            start: parseInt(get(
-                                                                listComponent,
-                                                                'rsConfig.startValue',
-                                                                ''
-                                                            ), 10),
-                                                            end: parseInt(get(
-                                                                listComponent,
-                                                                'rsConfig.endValue',
-                                                                ''
-                                                            ), 10),
-                                                        }}
-                                                        rangeLabels={{
-                                                            start: get(
-                                                                listComponent,
-                                                                'rsConfig.startLabel',
-                                                                ''
-                                                            ),
-                                                            end: get(
-                                                                listComponent,
-                                                                'rsConfig.endLabel',
-                                                                ''
-                                                            ),
-                                                        }}
-                                                        showHistogram={get(
-                                                            listComponent,
-                                                            'rsConfig.showHistogram',
-                                                            false
-                                                        )}
-                                                        URLParams
-                                                        css={this.getFontFamily()}
-                                                        filterLabel={get(
-                                                                listComponent,
-                                                                'rsConfig.filterLabel',
-                                                                ''
-                                                            ) || get(
-                                                                listComponent,
-                                                                'rsConfig.title',
-                                                                ''
-                                                            )
-                                                        }
-                                                    />
-                                                ) : (
-                                                    <DynamicRangeSlider
-                                                        key={get(
-                                                            listComponent,
-                                                            'rsConfig.componentId',
-                                                            ''
-                                                        )}
-                                                        componentId={get(
-                                                            listComponent,
-                                                            'rsConfig.componentId',
-                                                            ''
-                                                        )}
-                                                        dataField={get(
-                                                            listComponent,
-                                                            'rsConfig.dataField',
-                                                            ''
-                                                        )}
-                                                        showHistogram={get(
-                                                            listComponent,
-                                                            'rsConfig.showHistogram',
-                                                            false
-                                                        )}
-                                                        URLParams
-                                                        css={this.getFontFamily()}
-                                                        filterLabel={get(
-                                                            listComponent,
-                                                            'rsConfig.filterLabel',
-                                                            ''
-                                                            ) || get(
-                                                                listComponent,
-                                                                'rsConfig.title',
-                                                                ''
-                                                            )
-                                                        }
-                                                    />
-                                                )
-
-                                            ) : (
-                                                <MultiList
-                                                    key={get(
-                                                        listComponent,
-                                                        'rsConfig.componentId',
-                                                    )}
-                                                    innerClass={{
-                                                        input: 'list-input'
-                                                    }}
-                                                    componentId={get(
-                                                        listComponent,
-                                                        'rsConfig.componentId',
-                                                    )}
-                                                    URLParams
-                                                    loader={
-                                                        <div
-                                                            css={loaderStyle}
-                                                            // eslint-disable-next-line
-                                                            dangerouslySetInnerHTML={{
-                                                                __html: get(
-                                                                    listComponent,
-                                                                    'customMessages.loading',
-                                                                    'Loading options',
-                                                                ),
-                                                            }}
-                                                        />
-                                                    }
-                                                    renderNoResults={() => (
-                                                        <div
-                                                            // eslint-disable-next-line
-                                                            dangerouslySetInnerHTML={{
-                                                                __html: get(
-                                                                    listComponent,
-                                                                    'customMessages.noResults',
-                                                                    'No items Found',
-                                                                ),
-                                                            }}
-                                                        />
-                                                    )}
-                                                    showCount={
-                                                        this.themeType !== 'minimal'
-                                                    }
-                                                    showCheckbox={
-                                                        this.themeType !== 'minimal'
-                                                    }
-                                                    {...listComponent.rsConfig}
-                                                    dataField={get(
-                                                        listComponent,
-                                                        'rsConfig.dataField',
-                                                    )}
-                                                    css={this.getFontFamily()}
-                                                    react={{
-                                                        and: getReactDependenciesFromPreferences(
-                                                            this.preferences,
-                                                            get(
-                                                                listComponent,
-                                                                'rsConfig.componentId',
-                                                            ),
-                                                        ),
-                                                    }}
-                                                    filterLabel={get(
-                                                        listComponent,
-                                                        'rsConfig.filterLabel',
-                                                        ''
-                                                        ) || get(
-                                                            listComponent,
-                                                            'rsConfig.title',
-                                                            ''
-                                                        )
-                                                    }
-                                                    title=""
-                                                />
-                                            )
-                                        }
-                                    </Panel>
-                                ))}
-                            </Collapse>
-                        </div>
+                        <Filters
+                            theme={this.theme}
+                            isMobile={this.isMobile}
+                            themeType={this.themeType}
+                            preferences={this.preferences}
+                            dynamicFacets={this.dynamicFacets}
+                            getFontFamily={this.getFontFamily()}
+                            collectionFilter={this.collectionFilter}
+                            exportType={this.exportType}
+                            toggleFilters={toggleFilters}
+                            currency={this.currency}
+                        />
 
                         <div>
                             {this.themeType === 'minimal' &&
